Applied requested changes for pr
