+++ Generator.Equals/Extensions/SymbolExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
-
+using System.Diagnostics.CodeAnalysis;
+using Generator.Equals.Models;
 using Microsoft.CodeAnalysis;
 
 namespace Generator.Equals.Extensions;
@@ -38,5 +39,33 @@ internal static class SymbolExtensions
 
             yield return namespaceOrTypeSymbol;
         }
+    }
+    
+    [SuppressMessage("ReSharper", "ForeachCanBeConvertedToQueryUsingAnotherGetEnumerator", Justification = "Performance")]
+    public static bool HasAttribute(this ISymbol symbol, AttributeMetadata metadata)
+    {
+        foreach (var attribute in symbol.GetAttributes())
+        {
+            if (metadata.Equals(attribute.AttributeClass))
+            {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    //GetAttribute
+    public static AttributeData? GetAttribute(this ISymbol symbol, AttributeMetadata metadata)
+    {
+        foreach (var attribute in symbol.GetAttributes())
+        {
+            if (metadata.Equals(attribute.AttributeClass))
+            {
+                return attribute;
+            }
+        }
+
+        return null;
     }
 }
